server/requirements.txt
sqlalchemy
psycopg2
Scrapy
pytest-playwright
streamlit
pandas